fix(dashboard): preserve sandbox context values on autocomplete merge fixes NV-8340
Merge synthetic context fragments under existing preview data so selecting
another context field does not overwrite user-edited tenant ids. Also
dedupe context-type extraction and trim redundant comments.
