+++ apps/dashboard/src/hooks/use-parse-variables.ts
@@ -40,7 +40,6 @@ export function useParseVariables(
     const newVars = contextTypeVariables.filter((v) => !existingNames.has(v.name));
     if (newVars.length === 0) return result;
 
-    // Also land in primitives — Maily autocomplete only reads primitives/arrays/namespaces.
     return {
       ...result,
       variables: [...result.variables, ...newVars],
